up bai

+++ ss7_spring_data_repository/bai_tap/update_blog/src/main/java/com/example/blog/service/impl/BlogService.java
@@ -1,14 +1,12 @@
 package com.example.blog.service.impl;
 
 import com.example.blog.entity.Blog;
+import com.example.blog.exception.BlogNotFoundException;
 import com.example.blog.repository.IBlogRepository;
 import com.example.blog.service.IBlogService;
-import com.example.blog.service.IService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,11 +23,6 @@ public class BlogService implements IBlogService {
         return (List<Blog>) blogRepository.findAll();
     }
 
-//    @Override
-//    public Page<Blog> findAll(Pageable pageable) {
-//        return blogRepository.findAll(pageable);
-//    }
-
     @Override
     public void save(Blog b) {
         blogRepository.save(b);
@@ -37,7 +30,12 @@ public class BlogService implements IBlogService {
 
     @Override
     public void delete(Integer id) {
-        blogRepository.deleteById(id);
+        Optional<Blog> existingBlog = blogRepository.findById(id);
+        if (existingBlog.isPresent()) {
+            blogRepository.deleteById(id);
+        } else {
+            throw new BlogNotFoundException("Blog with ID " + id + " not found");
+        }
     }
 
     @Override
@@ -53,7 +51,11 @@ public class BlogService implements IBlogService {
             updatedBlog.setName(blog.getName());
             updatedBlog.setBlogger(blog.getBlogger());
             updatedBlog.setUrl(blog.getUrl());
+            updatedBlog.setCreatedAt(blog.getCreatedAt());
+            updatedBlog.setCategory(blog.getCategory());
             blogRepository.save(updatedBlog);
+        }else {
+            throw new BlogNotFoundException("Blog with ID " + id + " not found");
         }
     }
 
@@ -63,7 +65,13 @@ public class BlogService implements IBlogService {
     }
 
     @Override
-    public Slice<Blog> findAll(int page, int size) {
-        return blogRepository.findAll(PageRequest.of(page, size));
+    public Page<Blog> findAll(Pageable pageable) {
+        return blogRepository.findAll(pageable);
     }
+
+    @Override
+    public Page<Blog> findByNameContaining(String name, Pageable pageable) {
+        return blogRepository.findByNameContaining(name, pageable);
+    }
+
 }
